Move address step into form so progress can be more easily tracked
